+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 * discovery: derive "discovery_types" enum from all configured discovery types, across all envoy versions
 * ui: added a count for all resource types
 * ui: added support for selecting which envoy version templates should be rendered as
+* ui: added support for filtering resources by service cluster/source match key
 * ui: '/' now redirects to the UI instead of docs